Move MediatemsMapper to its own file
