+++ MovieListFramework/Media API/RemoteMediaLoader.swift
@@ -43,56 +43,3 @@ public final class RemoteMediaLoader {
     }
 }
 
-private class MediaItemsMapper {
-    private struct Root: Decodable {
-        let results: [Item]
-    }
-
-    private struct Item: Decodable {
-        let adult: Bool
-        let backdrop_path: String?
-        let genre_ids: [Int]
-        let id: UUID
-        let media_type: String?
-        let original_language: String
-        let original_title: String
-        let overview: String
-        let popularity: Double
-        let poster_path: String?
-        let release_date: String
-        let title: String
-        let video: Bool
-        let vote_average: Double
-        let vote_count: Int
-
-        var item: MediaItem {
-            return MediaItem(adult: adult,
-                             backdropPath: backdrop_path,
-                             genreIds: genre_ids,
-                             id: id,
-                             mediaType: media_type,
-                             originalLanguage: original_language,
-                             originalTitle: original_title,
-                             overview: overview,
-                             popularity: popularity,
-                             posterPath: poster_path,
-                             releaseDate: release_date,
-                             title: title,
-                             video: video,
-                             voteAverage: vote_average,
-                             voteCount: vote_count)
-        }
-    }
-    
-    static var OK_200: Int { return 200 }
-    
-    static func map(_ data: Data, _ response: HTTPURLResponse) throws -> [MediaItem] {
-        guard response.statusCode == OK_200 else {
-            throw RemoteMediaLoader.Error.invalidData
-        }
-        let root = try JSONDecoder().decode(Root.self, from: data)
-        return root.results.map { $0.item }
-    }
-}
-
-
